Update proxy to use tokio-proto
The proxy now uses the pipeline functionality from tokio-proto. The
proxy is now a tokio-service that combines a pipeline server and a
pipeline client. The pipeline server listens for requests from a
downstream client. The request is then passed to the pipeline client,
which will issue that same request upstream to a address from the pool.
The intention is to add HTTP pipeline support but also to make the
proxying of the request and response compeltely async.

Currently, the proxy will only accept requests without a body. I have
tested a few scenarios, but I am sure there are many bugs still. Also,
no performance testing has been done. I am sure there are allocations
being made that are not necessary.

The design of tokio-proto is changing from using a Transport trait to
using a Sink + Stream. I am saving my progress here so I can start to
incorporate those new changes. I believe the "interface" of the proxy
will not change much, which will allow me to also start adding features
and functionality that are unrelated to the implementation of
tokio-proto.

Other changes:

- minimized the amount of http parsing being done
- updated test-server to be an HTTP 1.1 compliant server