Use only a single thread
The proxy, management and health check streams are merged together so
they can all run on the same thread. This removes the overhead of having
multiple tokio Cores running at the same time and gets rid of dedicated
threads for tasks that were not often used. Most importantly, it
simplifies the Pool implementation to not require synchronization for
reading servers out.

I have also removed the dependency on rustful. Using rustful was
requiring two different versions of hyper, which was causing some odd
conflicts with dependencies.